ISO-related stuff in formatters-file
