+++ frontend/app/utils/formatters.ts
@@ -1,5 +1,71 @@
 // Formatting helpers used across the frontend (dates and ranges).
 
+// ─── ISO / date helpers ──────────────────────────────────────────────────────
+
+/** Serialize a Date to a local-timezone YYYY-MM-DD string (no UTC drift). */
+export function localIso(d: Date): string {
+  const y  = d.getFullYear();
+  const m  = String(d.getMonth() + 1).padStart(2, "0");
+  const dd = String(d.getDate()).padStart(2, "0");
+  return `${y}-${m}-${dd}`;
+}
+
+/** Return today's date as a local-timezone ISO string. */
+export function localTodayIso(): string {
+  return localIso(new Date());
+}
+
+/**
+ * Validate a YYYY-MM-DD string: checks calendar validity and rejects future dates.
+ * Returns the ISO string on success, null on failure.
+ */
+export function validateIso(iso: string): string | null {
+  if (!/^\d{4}-\d{2}-\d{2}$/.test(iso)) return null;
+  const [y, m, d] = iso.split("-").map(Number) as [number, number, number];
+  const date = new Date(y, m - 1, d);
+  if (
+    date.getFullYear() !== y ||
+    date.getMonth()    !== m - 1 ||
+    date.getDate()     !== d
+  ) return null;
+  if (iso > localTodayIso()) return null;
+  return iso;
+}
+
+/**
+ * Parse a user-typed date string in DD/MM/YYYY or YYYY-MM-DD format.
+ * Returns a validated ISO string, or null if unparseable / invalid.
+ */
+export function parseDate(raw: string): string | null {
+  const val = raw.replace(/[^\d\/\-\.]/g, "").trim();
+  if (/^\d{4}-\d{2}-\d{2}$/.test(val)) return validateIso(val);
+  const parts = val.split(/[\/\-\.]/);
+  if (parts.length === 3) {
+    const [dd, mm, yyyy] = parts;
+    if (yyyy?.length === 4 && dd && mm)
+      return validateIso(`${yyyy}-${mm.padStart(2, "0")}-${dd.padStart(2, "0")}`);
+  }
+  return null;
+}
+
+/** Format an ISO string as DD/MM/YYYY for display in text inputs. */
+export function formatForInput(iso: string): string {
+  const [y, m, d] = iso.split("-");
+  return `${d}/${m}/${y}`;
+}
+
+// ─── Locale / label helpers ──────────────────────────────────────────────────
+
+/** Produce 7 narrow weekday labels (Mon → Sun) for a given locale. */
+export function buildWeekdayLabels(intlLocale: string): string[] {
+  const mon = new Date(2024, 0, 1); // 1 Jan 2024 is a known Monday
+  return Array.from({ length: 7 }, (_, i) => {
+    const d = new Date(mon);
+    d.setDate(1 + i);
+    return d.toLocaleDateString(intlLocale, { weekday: "narrow" });
+  });
+}
+
 // Format ISO date to short string according to locale ('nl' or 'en')
 export function formatDateShort(
   iso: string,
